% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datagen.R
\name{datagen}
\alias{datagen}
\title{Generate example data}
\usage{
datagen(N, censor = 80)
}
\arguments{
\item{N}{integer.
The number of individuals in the dataset}

\item{censor}{numeric. The total observation period.}
}
\description{
Generate a data table with example data
}
\details{
The dataset simulates a labour market program. People entering the dataset are without a job,
there are two transitions, either to \code{"exit"}, in which case the person leaves the dataset,
or \code{"treatment"}. Those who have received treatment stays in the dataset, but may not
receive a new treatment. This process continues until either the person exits, or the
\code{"censor"} time is reached. There are two covariates, \code{"x1"} and \code{"x2"}, each with
their own effect on the hazards for the transitions. In addition there is some hidden
individual heterogeneity. It is this hidden heterogeneity which is modeled by the mixed proportional
hazard distribution.
}
\examples{
dataset <- datagen(4000,1200)
print(dataset)
risksets <- list(untreated=c(1,2), treated=c(1))
# just two iterations to save time
opt <- mphcrm(d ~ x1+x2|alpha, data=dataset, id=id, durvar=duration,state=alpha+1,risksets=risksets,
              control=mphcrm.control(threads=1,iters=2))
best <- opt[[1]]
print(best)
summary(best)
}
