% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{geninv}
\alias{geninv}
\title{Moore-Penrose generalized inverse}
\usage{
geninv(X, tol = sqrt(.Machine$double.eps))
}
\arguments{
\item{X}{matrix}

\item{tol}{tolerance for determining bad entries}
}
\description{
Moore-Penrose generalized inverse
}
