% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duration.R
\name{mphcrm.control}
\alias{mphcrm.control}
\title{Control parameters for mphcrm}
\usage{
mphcrm.control(...)
}
\arguments{
\item{...}{parameters to change}
}
\value{
list of control parameters suitable for the \code{control} argument of \code{\link{mphcrm}}
}
\description{
Control parameters for mphcrm
}
