% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duration.R
\name{mphcrm}
\alias{mphcrm}
\title{Estimate a mixed proportional hazard model}
\usage{
mphcrm(formula, data, id, durvar, state, risksets = NULL,
  timing = c("exact", "interval", "logit"), subset, na.action,
  control = mphcrm.control(), cluster = NULL)
}
\arguments{
\item{formula}{A formula specifying the covariates.
In a formula like \code{d ~ x1 + x2},
the \code{d} is the transition which is taken, coded as an integer where \code{0} means
no transition, and otherwise \code{d} is the number of the transition which is taken.
\code{d} can also be a factor, in which case the level which is no transition must be named
\code{"0"} or \code{"none"}.

The \code{x1+x2} part is like in \code{\link{lm}}, i.e. ordinary covariates or factors.

If the covariates differ among the transitions, a multi-part formula can be specified.
E.g. \code{d ~ x1+x2 | x3+x4 | x5+x6}, in which case the first part (\code{x1+x2}) is common
for all the transitions, the second part, \code{x3+x4}, is specific to the first transition,
the third part is specific for the second transition, and so on. I.e. the covariates
for transition 1 are \code{x1+x2+x3+x4}, whereas the covariates for transition 2 are \code{x1+x2+x5+x6}.}

\item{data}{A data frame which contains the covariates.}

\item{id}{The name of the covariate which identifies an individual.}

\item{durvar}{The name of the covariate which contains the length of the observation period. Can be a
numeric, in which case all observations are assumed to have the same length. If missing, it
is assumed to be 1.}

\item{state}{The name of a variable which is an index into \code{risksets}, specifying for each observation
which risks are present.}

\item{risksets}{A list of integer vectors. Each vector is a list of transitions, which risks are present for
the observation.}

\item{subset}{For specifying a subset of the dataset, similar to \code{\link{lm}}.}

\item{na.action}{For handling of missing cases, similar to \code{\link{lm}}.}

\item{control}{List of control parameters for the estimation. See \code{\link{mphcrm.control}}.}

\item{cluster}{A cluster specification from package \pkg{parallel} or \pkg{snow}. There is quite some
overhead in running on a cluster, primarily intended for very large datasets. If you have
more than one cpu on a machine, use \code{control=\link{mphcrm.control}(threads=n)} instead.}
}
\value{
A list, one entry for each iteration. Ordered in reverse order. Ordinarily you will be
interested in the first entry.
}
\description{
\code{mphcrm} implements estimation of a mixed proportional hazard competing risk model.
The baseline hazard is of the form \eqn{exp(X \beta)} where \eqn{X} is a
matrix of covariates, and \eqn{\beta} is a vector of paramaters to estimate.
In addition there is an intercept term \eqn{\mu}, i.e. the hazard is \eqn{exp(X \beta + \mu)}.
There are several transitions to be made, and a set of (\eqn{X}, \eqn{\beta}, and \eqn{\mu}) for
each possible transition.

Each individual may have several observations, with either a transition at the end of the
observation, or not a transition. It is a competing risk, there can be more than one possible
transition for an observation, but only one is taken at the end of the period.

For each individual \eqn{i} there is a log likelihood as a function of \eqn{\mu}, called \eqn{M_i(\mu)}.

The mixture is that the \eqn{\mu}'s are stochastic. I.e. 
we have probabilities \eqn{p_j}, and a vector of \eqn{\mu_j} of masspoints (one for each transition),
for each such \eqn{j}.

So the full likelihood for an individual is \eqn{L_i = \sum_j p_j M_i(\mu_j)}.

The \code{mphcrm()} function maximizes the likelihood \eqn{\sum_i L_i} over
\eqn{p_j}, \eqn{mu_j}, and \eqn{\beta}.

In addition to the parameters \eqn{\beta}, a variable which records the duration of each
observation must be specified.

In some datasets it is known that not all risks are present at all times. Like, losing your
job when you do not have one. In this case it should be specified which risks are present.

The estimation starts out with one masspoint, maximizes the likelihood, tries to add another
point, and continues in this fashion.
}
\details{
The estimation starts by estimating the null-model, i.e. all parameters set to 0, only one
intercept for each transition is estimated.

Then it estimates the full model, still with one intercept in each transition.

After the initial model has been estimated, it tries to add a masspoint to the mixing
distribution, then estimates the model with this new distribution.

It continues to add masspoints in this way until eithr it can not improve the likelihood, or
the number of iterations as specified in control$iters are reached.

The result of every iteration is returned in a list.

If you interrupt \code{mphcrm} it will catch the interrupt and return with the
estimates it has found so far.
}
