% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mphdist}
\alias{mphdist}
\alias{mphdist.log}
\alias{mphmoments}
\alias{mphmoments.log}
\title{Extract the mixed proportional hazard distribution}
\usage{
mphdist(pset)

mphdist.log(pset)

mphmoments(pset)

mphmoments.log(pset)
}
\arguments{
\item{pset}{a parameter set of class \code{"mphcrm.pset"}, typically
\code{opt[[1]]$par}, where \code{opt} is returned from \code{\link{mphcrm}}.}
}
\value{
matrix where there is one row for each masspoint. The first consists of the probabilities,
the other columns are the hazards for each transition.
}
\description{


Same as \code{mphdist}, but the logarithms of the hazards.

Extract moments of the distribution

Same as \code{mphmoments}, but for log hazards
}
