% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mphmoments}
\alias{mphmoments}
\alias{mphmoments.log}
\title{Extract moments of the mixed proportional hazard distribution}
\usage{
mphmoments(pset, round = 20)

mphmoments.log(pset)
}
\description{
Extract moments of the mixed proportional hazard distribution
}
\seealso{
@mphdist
}
