% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{se}
\alias{se}
\title{Extract standard errors of the estimated parameters}
\usage{
se(x, tol = 1e-09)
}
\arguments{
\item{x}{The Fisher matrix, typically from \code{opt[[1]]$fisher}, where \code{opt} is returned
from \code{\link{mphcrm}}.}
}
\description{
Extract standard errors of the estimated parameters
}
