% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mphdist}
\alias{mphdist}
\title{Extract the mixed proportional hazard distribution}
\usage{
mphdist(pset, round = 20)
}
\arguments{
\item{pset}{a parameter set of class \code{"mphcrm.pset"}, typically
\code{opt[[1]]$par}, where \code{opt} is returned from \code{\link{mphcrm}}.}

\item{round}{integer. Number of decimals. In a competing risk model, some of the hazards can be very small, \code{mphdist}
can round the value so that e.q. \code{1e-8} becomes zero, mostly useful for nice displays.}
}
\value{
matrix where there is one row for each masspoint. The first consists of the probabilities,
the other columns are the hazards for each transition.
}
\description{

}
