% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{se}
\alias{se}
\title{Extract standard errors of the estimated parameters}
\usage{
se(x, tol = .Machine$double.eps)
}
\arguments{
\item{x}{The Fisher matrix, typically from \code{opt[[1]]$fisher}, where \code{opt} is returned
from \code{\link{mphcrm}}.}

\item{tol}{tolerance for \link{geninv}}
}
\value{
A named vector of standard errors is returned.
}
\description{
Extract standard errors of the estimated parameters
}
