% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{timestr}
\alias{timestr}
\title{Prettyprint a time interval}
\usage{
timestr(t)
}
\arguments{
\item{t}{numeric. time in seconds.}
}
\value{
A character string is returned.
}
\description{
Converts a time in seconds to a short string e.g. \code{"3m4s"}.
}
\examples{
timestr(1.3)
timestr(73)
timestr(4684)
}
